+++ bot/.pylintrc
@@ -209,10 +209,6 @@ disable=print-statement,
         keyword-arg-before-vararg,
         # not smart enough about refinement
         inconsistent-return-statements,
-        # doesn't understand modern dict
-        unsubscriptable-object,
-        # mypy handles this
-        no-value-for-parameter,
 
 
 # Enable the message, report, category or checker with the given id(s). You can
