+++ bot/kodiak/events/__init__.py
@@ -1,12 +1,16 @@
-import pydantic
+from typing import Mapping, Type
 
-from kodiak.events.check_run import CheckRunEvent  # noqa: F401
-from kodiak.events.pull_request import PullRequestEvent  # noqa: F401
-from kodiak.events.pull_request_review import PullRequestReviewEvent  # noqa: F401
-from kodiak.events.push import PushEvent  # noqa: F401
-from kodiak.events.status import StatusEvent  # noqa: F401
+from kodiak.events.base import GithubEvent
+from kodiak.events.check_run import CheckRunEvent
+from kodiak.events.pull_request import PullRequestEvent
+from kodiak.events.pull_request_review import PullRequestReviewEvent
+from kodiak.events.push import PushEvent
+from kodiak.events.status import StatusEvent
 
-
-class RawIncomingEvent(pydantic.BaseModel):
-    name: str
-    payload: str
+event_schema_mapping: Mapping[str, Type[GithubEvent]] = {
+    "check_run": CheckRunEvent,
+    "pull_request": PullRequestEvent,
+    "pull_request_review": PullRequestReviewEvent,
+    "push": PushEvent,
+    "status": StatusEvent,
+}
