+++ bot/kodiak/main.py
@@ -6,21 +6,19 @@ import logging
 import sys
 from typing import Any, Dict, Optional, cast
 
-import asyncio_redis
 import sentry_sdk
 import structlog
 import uvicorn
 from fastapi import FastAPI, Header, HTTPException
-from fastapi.param_functions import Depends
 from sentry_sdk.integrations.asgi import SentryAsgiMiddleware
 from sentry_sdk.integrations.logging import LoggingIntegration
 from starlette import status
 from starlette.requests import Request
 
 from kodiak import app_config as conf
-from kodiak import redis
+from kodiak.event_handlers import handle_webhook_event
 from kodiak.logging import SentryProcessor, add_request_info_processor
-from kodiak.queue import enqueue_incoming_webhook
+from kodiak.queue import redis_webhook_queue
 
 # for info on logging formats see: https://docs.python.org/3/library/logging.html#logrecord-attributes
 logging.basicConfig(
@@ -70,7 +68,6 @@ async def webhook_event(
     request: Request,
     x_github_event: str = Header(None),
     x_hub_signature: str = Header(None),
-    redis: asyncio_redis.Connection = Depends(redis.get_conn),
 ) -> None:
     """
     Verify and accept GitHub Webhook payloads.
@@ -100,15 +97,13 @@ async def webhook_event(
             detail="Invalid signature: X-Hub-Signature",
         )
 
-    await enqueue_incoming_webhook(redis=redis, event_name=github_event, event=event)
+    await handle_webhook_event(event_name=github_event, payload=event)
 
 
 @app.on_event("startup")
 async def startup() -> None:
-    # create redis queue so the first request to the HTTP server doesn't have to
-    # wait for the queue creation.
-    await redis.get_conn()
+    await redis_webhook_queue.create()
 
 
-def main() -> None:
+if __name__ == "__main__":
     uvicorn.run("kodiak.main:app", host="0.0.0.0", port=conf.PORT)
