+++ bot/kodiak/queue.py
@@ -1,35 +1,21 @@
 from __future__ import annotations
 
 import asyncio
-import json
 import time
 import typing
 import urllib
 from datetime import timedelta
-from typing import Iterator, NoReturn
+from typing import Optional
 
 import asyncio_redis
 import sentry_sdk
 import structlog
-import zstandard as zstd
 from asyncio_redis import Pool as RedisConnection
 from asyncio_redis.replies import BlockingZPopReply
 from pydantic import BaseModel
-from typing_extensions import Protocol
 
-from kodiak import app_config as conf
-from kodiak import queries
-from kodiak.events import (
-    CheckRunEvent,
-    PullRequestEvent,
-    PullRequestReviewEvent,
-    PushEvent,
-    RawIncomingEvent,
-    StatusEvent,
-)
-from kodiak.events.status import Branch
+import kodiak.app_config as conf
 from kodiak.pull_request import evaluate_pr
-from kodiak.queries import Client
 
 logger = structlog.get_logger()
 
@@ -40,231 +26,6 @@ WEBHOOK_QUEUE_NAMES = "kodiak_webhook_queue_names"
 WORKER_TASKS: typing.MutableMapping[str, asyncio.Task[None]] = {}
 
 RETRY_RATE_SECONDS = 2
-
-
-class WebhookQueueProtocol(Protocol):
-    async def enqueue(self, *, event: WebhookEvent) -> None:
-        ...
-
-    async def enqueue_for_repo(self, *, event: WebhookEvent, first: bool) -> int | None:
-        ...
-
-
-async def pr_event(queue: WebhookQueueProtocol, pr: PullRequestEvent) -> None:
-    """
-    Trigger evaluation of modified PR.
-    """
-    await queue.enqueue(
-        event=WebhookEvent(
-            repo_owner=pr.repository.owner.login,
-            repo_name=pr.repository.name,
-            pull_request_number=pr.number,
-            target_name=pr.pull_request.base.ref,
-            installation_id=str(pr.installation.id),
-        )
-    )
-
-
-def check_run(check_run_event: CheckRunEvent) -> Iterator[WebhookEvent]:
-    """
-    Trigger evaluation of all PRs included in check run.
-    """
-    # Prevent an infinite loop when we update our check run
-    if check_run_event.check_run.name == queries.CHECK_RUN_NAME:
-        return
-    for pr in check_run_event.check_run.pull_requests:
-        # filter out pull requests for other repositories
-        if pr.base.repo.id != check_run_event.repository.id:
-            continue
-        yield WebhookEvent(
-            repo_owner=check_run_event.repository.owner.login,
-            repo_name=check_run_event.repository.name,
-            pull_request_number=pr.number,
-            target_name=pr.base.ref,
-            installation_id=str(check_run_event.installation.id),
-        )
-
-
-def find_branch_names_latest(sha: str, branches: list[Branch]) -> list[str]:
-    """
-    from the docs:
-        The "branches" key is "an array of branch objects containing the status'
-        SHA. Each branch contains the given SHA, but the SHA may or may not be
-        the head of the branch. The array includes a maximum of 10 branches.""
-    https://developer.github.com/v3/activity/events/types/#statusevent
-
-    NOTE(chdsbd): only take branches with commit at branch head to reduce
-    potential number of api requests we need to make.
-    """
-    return [branch.name for branch in branches if branch.commit.sha == sha]
-
-
-async def status_event(queue: WebhookQueueProtocol, status_event: StatusEvent) -> None:
-    """
-    Trigger evaluation of all PRs associated with the status event commit SHA.
-    """
-    owner = status_event.repository.owner.login
-    repo = status_event.repository.name
-    installation_id = str(status_event.installation.id)
-    log = logger.bind(owner=owner, repo=repo, install=installation_id)
-
-    refs = find_branch_names_latest(
-        sha=status_event.sha, branches=status_event.branches
-    )
-
-    async with Client(
-        owner=owner, repo=repo, installation_id=installation_id
-    ) as api_client:
-        if len(refs) == 0:
-            # when a pull request is from a fork the status event will not have
-            # any `branches`, so to be able to trigger evaluation of the PR, we
-            # fetch all pull requests.
-            #
-            # I think we could optimize this by selecting only the fork PRs, but
-            # I worry that we might miss some events where `branches` is empty,
-            # but not because of a fork.
-            pr_results = [await api_client.get_open_pull_requests()]
-            log.info("could not find refs for status_event")
-        else:
-            pr_requests = [
-                api_client.get_open_pull_requests(head=f"{owner}:{ref}") for ref in refs
-            ]
-            pr_results = await asyncio.gather(*pr_requests)
-
-        all_events: set[WebhookEvent] = set()
-        for prs in pr_results:
-            if prs is None:
-                continue
-            for pr in prs:
-                all_events.add(
-                    WebhookEvent(
-                        repo_owner=owner,
-                        repo_name=repo,
-                        pull_request_number=pr.number,
-                        target_name=pr.base.ref,
-                        installation_id=str(installation_id),
-                    )
-                )
-        for event in all_events:
-            await queue.enqueue(event=event)
-
-
-async def pr_review(
-    queue: WebhookQueueProtocol, review: PullRequestReviewEvent
-) -> None:
-    """
-    Trigger evaluation of the modified PR.
-    """
-    await queue.enqueue(
-        event=WebhookEvent(
-            repo_owner=review.repository.owner.login,
-            repo_name=review.repository.name,
-            pull_request_number=review.pull_request.number,
-            target_name=review.pull_request.base.ref,
-            installation_id=str(review.installation.id),
-        )
-    )
-
-
-def get_branch_name(raw_ref: str) -> str | None:
-    """
-    Extract the branch name from the ref
-    """
-    if raw_ref.startswith("refs/heads/"):
-        return raw_ref.split("refs/heads/", 1)[1]
-    return None
-
-
-async def push(queue: WebhookQueueProtocol, push_event: PushEvent) -> None:
-    """
-    Trigger evaluation of PRs that depend on the pushed branch.
-    """
-    owner = push_event.repository.owner.login
-    repo = push_event.repository.name
-    installation_id = str(push_event.installation.id)
-    branch_name = get_branch_name(push_event.ref)
-    log = logger.bind(ref=push_event.ref, branch_name=branch_name)
-    if branch_name is None:
-        log.info("could not extract branch name from ref")
-        return
-    async with Client(
-        owner=owner, repo=repo, installation_id=installation_id
-    ) as api_client:
-        # find all the PRs that depend on the branch affected by this push and
-        # queue them for evaluation.
-        # Any PR that has a base ref matching our event ref is dependent.
-        prs = await api_client.get_open_pull_requests(base=branch_name)
-        if prs is None:
-            log.info("api call to find pull requests failed")
-            return None
-        for pr in prs:
-            await queue.enqueue(
-                event=WebhookEvent(
-                    repo_owner=owner,
-                    repo_name=repo,
-                    pull_request_number=pr.number,
-                    target_name=pr.base.ref,
-                    installation_id=installation_id,
-                )
-            )
-
-
-_redis: asyncio_redis.Pool | None = None
-
-
-async def get_redis() -> asyncio_redis.Pool:
-    global _redis  # pylint: disable=global-statement
-    if _redis is None:
-        _redis = await asyncio_redis.Pool.create(
-            host=conf.REDIS_URL.hostname or "localhost",
-            port=conf.REDIS_URL.port or 6379,
-            password=(
-                conf.REDIS_URL.password.encode() if conf.REDIS_URL.password else None
-            ),
-            poolsize=conf.USAGE_REPORTING_POOL_SIZE,
-            encoder=asyncio_redis.encoders.BytesEncoder(),
-        )
-    return _redis
-
-
-def compress_payload(data: dict[str, object]) -> bytes:
-    cctx = zstd.ZstdCompressor()
-    return cctx.compress(json.dumps(data).encode())
-
-
-async def handle_webhook_event(
-    queue: WebhookQueueProtocol, event_name: str, payload: dict[str, object]
-) -> None:
-    log = logger.bind(event_name=event_name)
-
-    if conf.USAGE_REPORTING and event_name in conf.USAGE_REPORTING_EVENTS:
-        # store events in Redis for dequeue by web api job.
-        #
-        # We limit the queue length to ensure that if the dequeue job fails, we
-        # won't overload Redis.
-        redis = await get_redis()
-        await redis.rpush(
-            b"kodiak:webhook_event",
-            [compress_payload(dict(event_name=event_name, payload=payload))],
-        )
-        await redis.ltrim(b"kodiak:webhook_event", 0, conf.USAGE_REPORTING_QUEUE_LENGTH)
-        log = log.bind(usage_reported=True)
-
-    if event_name == "check_run":
-        for event in check_run(CheckRunEvent.parse_obj(payload)):
-            await queue.enqueue(event=event)
-    elif event_name == "pull_request":
-        await pr_event(queue, PullRequestEvent.parse_obj(payload))
-    elif event_name == "pull_request_review":
-        await pr_review(queue, PullRequestReviewEvent.parse_obj(payload))
-    elif event_name == "push":
-        await push(queue, PushEvent.parse_obj(payload))
-    elif event_name == "status":
-        await status_event(queue, StatusEvent.parse_obj(payload))
-    else:
-        log = log.bind(event_parsed=False)
-
-    log.info("webhook_event_handled")
 
 
 class WebhookEvent(BaseModel):
@@ -294,7 +55,7 @@ class WebhookEvent(BaseModel):
 
 async def process_webhook_event(
     connection: RedisConnection,
-    webhook_queue: WebhookQueueProtocol,
+    webhook_queue: RedisWebhookQueue,
     queue_name: str,
     log: structlog.BoundLogger,
 ) -> None:
@@ -319,7 +80,7 @@ async def process_webhook_event(
             only_if_not_exists=True,
         )
 
-    async def queue_for_merge(*, first: bool) -> int | None:
+    async def queue_for_merge(*, first: bool) -> Optional[int]:
         return await webhook_queue.enqueue_for_repo(event=webhook_event, first=first)
 
     log.info("evaluate pr for webhook event")
@@ -337,7 +98,7 @@ async def process_webhook_event(
 
 
 async def webhook_event_consumer(
-    *, connection: RedisConnection, webhook_queue: WebhookQueueProtocol, queue_name: str
+    *, connection: RedisConnection, webhook_queue: RedisWebhookQueue, queue_name: str
 ) -> typing.NoReturn:
     """
     Worker to process incoming webhook events from redis
@@ -360,18 +121,6 @@ async def webhook_event_consumer(
         log.info("start webhook event consumer")
         while True:
             await process_webhook_event(connection, webhook_queue, queue_name, log)
-
-
-async def process_raw_webhook_event_consumer(
-    *, queue: WebhookQueueProtocol, connection: RedisConnection
-) -> None:
-    res = await connection.blpop([INCOMING_QUEUE_NAME])
-
-    raw_event = RawIncomingEvent.parse_raw(res.value)
-    event_name = raw_event.name
-    event = json.loads(raw_event.payload)
-
-    await handle_webhook_event(queue=queue, event_name=event_name, payload=event)
 
 
 async def process_repo_queue(
@@ -397,7 +146,7 @@ async def process_repo_queue(
             only_if_not_exists=True,
         )
 
-    async def queue_for_merge(*, first: bool) -> int | None:
+    async def queue_for_merge(*, first: bool) -> Optional[int]:
         raise NotImplementedError
 
     log.info("evaluate PR for merging")
@@ -440,7 +189,7 @@ async def repo_queue_consumer(
 T = typing.TypeVar("T")
 
 
-def find_position(x: typing.Iterable[T], v: T) -> int | None:
+def find_position(x: typing.Iterable[T], v: T) -> typing.Optional[int]:
     count = 0
     for item in x:
         if item == v:
@@ -469,8 +218,6 @@ class RedisWebhookQueue:
             poolsize=conf.REDIS_POOL_SIZE,
         )
 
-        self.start_raw_webhook_worker()
-
         # restart repo workers
         merge_queues, webhook_queues = await asyncio.gather(
             self.connection.smembers(MERGE_QUEUE_NAMES),
@@ -483,21 +230,6 @@ class RedisWebhookQueue:
         for webhook_result in webhook_queues:
             queue_name = await webhook_result
             self.start_webhook_worker(queue_name=queue_name)
-
-    def start_raw_webhook_worker(self) -> None:
-        async def raw_webhook_event_consumer(
-            *, queue: WebhookQueueProtocol, connection: RedisConnection
-        ) -> NoReturn:
-            logger.info("start raw webhook event consumer")
-            while True:
-                await process_raw_webhook_event_consumer(
-                    queue=queue, connection=connection
-                )
-
-        self._start_worker(
-            INCOMING_QUEUE_NAME,
-            raw_webhook_event_consumer(queue=self, connection=self.connection),
-        )
 
     def start_webhook_worker(self, *, queue_name: str) -> None:
         self._start_worker(
@@ -547,7 +279,9 @@ class RedisWebhookQueue:
         log.info("enqueue webhook event")
         self.start_webhook_worker(queue_name=queue_name)
 
-    async def enqueue_for_repo(self, *, event: WebhookEvent, first: bool) -> int | None:
+    async def enqueue_for_repo(
+        self, *, event: WebhookEvent, first: bool
+    ) -> Optional[int]:
         """
         1. get the corresponding repo queue for event
         2. add key to MERGE_QUEUE_NAMES so on restart we can recreate the
@@ -600,17 +334,4 @@ def get_webhook_queue_name(event: WebhookEvent) -> str:
     return f"webhook:{event.installation_id}"
 
 
-INCOMING_QUEUE_NAME = "kodiak:incoming_queue"
-
-
-async def enqueue_incoming_webhook(
-    *, redis: asyncio_redis.Connection, event_name: str, event: dict[str, object]
-) -> None:
-    logger.info("enqueuing github webhook event")
-
-    raw_event = RawIncomingEvent(name=event_name, payload=json.dumps(event))
-    await redis.rpush(INCOMING_QUEUE_NAME, [raw_event.json()])
-
-
-def main() -> None:
-    asyncio.run(RedisWebhookQueue().create())
+redis_webhook_queue = RedisWebhookQueue()
